{00-SlitH-TB-01/00-SlitH-TB-01.yml, 02-ScraperH-SI-01SA/02-ScraperH-SI-01SA.yml}: Fix positive and negative coordinates
In the accelerator coordinate system, the position in the x axis increases in the direction of leaving the accelerator ring, while the position decreases when moving closer to the center of the machine.
Following this definition, the positive and negative edges of the Horizontal Slit and Scraper should be swapped.